% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibaq4diann-package.R
\docType{package}
\name{ibaq4diann-package}
\alias{ibaq4diann}
\alias{ibaq4diann-package}
\title{ibaq4diann: iBAQ Calculation from DIA-NN Proteomics Reports}
\description{
Computes intensity-Based Absolute Quantification (iBAQ) values from DIA-NN search results (report.parquet files) and a reference proteome FASTA. Implements the iBAQ method of Schwanhausser et al. (2011) \doi{10.1038/nature10098}, including in-silico tryptic digestion, quality filtering of DIA-NN reports, and flexible output formats. Also provides plotting utilities for visualising protein abundance ranks, distributions, sample correlations, and dynamic range.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlisonChaves/ibaq4diann}
  \item Report bugs at \url{https://github.com/AlisonChaves/ibaq4diann/issues}
}

}
\author{
\strong{Maintainer}: Alison Chaves \email{felipealison@gmail.com}

}
\keyword{internal}
